pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-lmstudio"
version = "0.0.1"
description = "These custom nodes acts as a bidirectional bridge between ComfyUI's synchronous/tensor-based ecosystem and LM Studio's asynchronous/text-based inference engine."
authors = [
  {name = "opj161", email = "opj161@outlook.com"}
]
readme = "README.md"
license = {text = "GNU General Public License v3"}
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "aiohttp",
    "Pillow",
    "numpy",
    "lmstudio",
    "comfyui-frontend-package"
]



[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
Repository = "https://github.com/opj161/comfyui-lmstudio"
BugTracker = "https://github.com/opj161/comfyui-lmstudio/issues"
Documentation = "https://github.com/opj161/comfyui-lmstudio/wiki"


[tool.comfy]
PublisherId = "opj161"
DisplayName = "comfyui-comfyui-lmstudio"
Icon = ""
Tags = []
Repository = "https://github.com/opj161/comfyui-lmstudio"

includes = []

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
